Move files from extensions directory to extension
