+++ floor_generator/lib/processor/field_processor.dart
@@ -3,8 +3,8 @@ import 'package:dartx/dartx.dart';
 import 'package:floor_annotation/floor_annotation.dart' as annotations;
 import 'package:floor_generator/misc/annotations.dart';
 import 'package:floor_generator/misc/constants.dart';
-import 'package:floor_generator/misc/extensions/type_converter_element_extension.dart';
-import 'package:floor_generator/misc/extensions/type_converters_extension.dart';
+import 'package:floor_generator/misc/extension/type_converter_element_extension.dart';
+import 'package:floor_generator/misc/extension/type_converters_extension.dart';
 import 'package:floor_generator/misc/type_utils.dart';
 import 'package:floor_generator/processor/processor.dart';
 import 'package:floor_generator/value_object/field.dart';
