+++ floor_generator/lib/value_object/database.dart
@@ -1,6 +1,6 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:collection/collection.dart';
-import 'package:floor_generator/misc/extensions/list_equality_extension.dart';
+import 'package:floor_generator/misc/extension/list_equality_extension.dart';
 import 'package:floor_generator/value_object/dao_getter.dart';
 import 'package:floor_generator/value_object/entity.dart';
 import 'package:floor_generator/value_object/type_converter.dart';
